better shutdown handling
